Render_upload/requirements.txt
# FastAPI web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==12.0

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.20

# Environment variables
python-dotenv==1.0.1

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Utilities
python-dateutil==2.9.0
loguru==0.7.3
